+++ pkg/analyzer/lib/src/dart/element/scope.dart
@@ -7,6 +7,7 @@ import 'package:analyzer/dart/ast/token.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/scope.dart';
 import 'package:analyzer/src/dart/ast/token.dart';
+import 'package:analyzer/src/dart/element/element.dart';
 import 'package:analyzer/src/dart/resolver/scope.dart' as impl;
 import 'package:meta/meta.dart';
 
@@ -24,5 +25,91 @@ class LibraryScope implements Scope {
     var token = SyntheticStringToken(TokenType.IDENTIFIER, name, 0);
     var identifier = astFactory.simpleIdentifier(token);
     return _implScope.lookup(identifier, _libraryElement);
+  }
+}
+
+class PrefixScope implements Scope {
+  final PrefixElement _element;
+  final Map<String, Element> _getters = {};
+  final Map<String, Element> _setters = {};
+
+  PrefixScope(this._element) {
+    for (var import in _element.enclosingElement.imports) {
+      if (import.prefix == _element) {
+        var elements = impl.NamespaceBuilder().getImportedElements(import);
+        elements.forEach(_add);
+      }
+    }
+  }
+
+  @override
+  Element lookup({@required String id, @required bool setter}) {
+    var map = setter ? _setters : _getters;
+    return map[id];
+  }
+
+  void _add(Element element) {
+    var setter = element is PropertyAccessorElement && element.isSetter;
+    _addTo(
+      map: setter ? _setters : _getters,
+      element: element,
+    );
+  }
+
+  void _addTo({
+    @required Map<String, Element> map,
+    @required Element element,
+  }) {
+    var id = element.displayName;
+
+    var existing = map[id];
+    if (existing != null && existing != element) {
+      map[id] = _merge(existing, element);
+      return;
+    }
+
+    map[id] = element;
+  }
+
+  Element _merge(Element existing, Element other) {
+    if (_isSdkElement(existing)) {
+      if (!_isSdkElement(other)) {
+        return other;
+      }
+    } else {
+      if (_isSdkElement(other)) {
+        return existing;
+      }
+    }
+
+    var conflictingElements = <Element>{};
+    _addElement(conflictingElements, existing);
+    _addElement(conflictingElements, other);
+
+    var definingLibrary = _element.enclosingElement;
+    return MultiplyDefinedElementImpl(
+      definingLibrary.context,
+      definingLibrary.session,
+      conflictingElements.first.name,
+      conflictingElements.toList(),
+    );
+  }
+
+  static void _addElement(
+    Set<Element> conflictingElements,
+    Element element,
+  ) {
+    if (element is MultiplyDefinedElementImpl) {
+      conflictingElements.addAll(element.conflictingElements);
+    } else {
+      conflictingElements.add(element);
+    }
+  }
+
+  static bool _isSdkElement(Element element) {
+    if (element is NeverElementImpl) {
+      return true;
+    }
+    return element.library.isInSdk;
   }
 }